always set initiator
