+++ packages/agent-core/src/lib/sessions.ts
@@ -7,30 +7,7 @@ import {
   paginateQuery,
 } from '@aws-sdk/lib-dynamodb';
 import { ddb, TableName } from './aws';
-
 import { AgentStatus, SessionItem } from '../schema';
-
-export const saveSessionInfo = async (workerId: string, initialMessage: string) => {
-  const now = Date.now();
-  const timestamp = String(now).padStart(15, '0');
-
-  await ddb.send(
-    new PutCommand({
-      TableName,
-      Item: {
-        PK: 'sessions',
-        SK: workerId,
-        workerId,
-        createdAt: now,
-        LSI1: timestamp,
-        initialMessage,
-        instanceStatus: 'terminated',
-        sessionCost: 0,
-        agentStatus: 'pending',
-      } satisfies SessionItem,
-    })
-  );
-};
 
 /**
  * Get session information from DynamoDB
